+++ test/test_utils.py
@@ -48,13 +48,17 @@ def test_phi_eq():
 
 
 def test_marginalize_out(s):
-    marginalized_distribution = utils.marginalize_out(s.nodes[0].index,
-                                                      s.network.tpm)
+    marginalized_distribution = utils.marginalize_out([0], s.tpm)
     assert np.array_equal(marginalized_distribution,
                           np.array([[[[0.,  0.,  0.5],
                                       [1.,  1.,  0.5]],
                                      [[1.,  0.,  0.5],
                                       [1.,  1.,  0.5]]]]))
+
+    marginalized_distribution = utils.marginalize_out([0, 1], s.tpm)
+    assert np.array_equal(marginalized_distribution,
+                          np.array([[[[ 0.5,  0. ,  0.5],
+                                      [ 1. ,  1. ,  0.5]]]]))
 
 
 def test_purview_max_entropy_distribution():